requirements.txt
# Audio capture
PyAudioWPatch>=0.2.12.8
numpy>=1.24.0
scipy>=1.10.0

# Transcription
deepgram-sdk>=6.0.0
websockets>=12.0

# PDF parsing
PyMuPDF>=1.23.0

# Knowledge base / RAG
openai>=1.30.0
faiss-cpu>=1.7.4
tiktoken>=0.5.0

# UI
PyQt6>=6.6.0

# Utilities
python-dotenv>=1.0.0
